requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pandas>=2.0.0
rdflib>=6.3.0
pystardog>=0.9.0
scikit-learn>=1.3.0
numpy>=1.24.0

